weather-etl-pipeline/requirements.txt
prefect>=2.0
requests>=2.28
pandas>=1.5
clickhouse-connect>=0.6
minio>=7.0
python-dotenv>=0.21
python-telegram-bot>=20.0